==> 05_Misc/Code Text mining-20191111/01_tweet_extraction.ipynb <==
# Tweet extraction 

Hello everyone! This notebook will discuss the `Twitter API`, `Tweepy`, and `TextBlob`. The `Tweepy` package is used to access the `Twitter API` and gather public tweets. We will also be using `TextBlob` to preprocess our tweets for later text analysis. 

import tweepy
import pandas as pd

### Twitter API
The `Twitter API` is an application programming interface that allows programmers to interact with Twitter. To access it, users are required to secure a `Twitter Developer Account` and a `Twitter App`.

Once both are secured, the user will have access to their assigned application keys. There are keys provided below. 

# Application keys
consumer_key = 'VpyTYjQp1p4RH5HLBESCQQYBe'
consumer_key_secret = 'ggKmMrIz5Hw7xmR4eskK93MoHvaX33kM6npShNGggP0XayNObo'

row1_consumer_key = 'ryxIRRxFIB3UE6DftqDVBFVIa'
row1_consumer_key_secret = '5Nn7bHCn3V6JNJyJlGZ1v4xKuXjnnSypRrYMCA0bmYZBj0IpMP'

row2_consumer_key = 'kqVi6r1TlPvZqdDwv0Gx0g8b0'
row2_consumer_key_secret = ''

row3_consumer_key = 'fkirrtsfNYN2y4WxYj1R1OsFG'
row3_consumer_key_secret = 'kxr1YhZ9SPtpi1ZCirMBkL4yCSZMEjTRWXw31bdJoLshtrhRua'

access_token = ''
access_token_secret = ''

def create_api(consumer_key, consumer_key_secret):
    """
    Creates an API instance with the provided access tokens
    """
    auth = tweepy.OAuthHandler(consumer_key, consumer_key_secret)
    auth.set_access_token(access_token, access_token_secret)
    api = tweepy.API(auth)
    return api

def search_tweets(api, query, ignore_rt=False, max_tweets=1000, lang='en'):
    """
    Searches public tweets given the API, query, and other optional parameters
    """
    if ignore_rt:
        query = query + ' -filter:retweets'
    search = tweepy.Cursor(api.search, 'apple', lang=lang).items(max_tweets)
    print("Downloaded {0} tweets".format(max_tweets))
    return search

def search_to_df(search_results):
    """
    Transform search results into a Pandas dataframe
    """
    tweets = [[t.id, t.user.screen_name, t.created_at, t.text] for t in search_results]
    tweets_df = pd.DataFrame(tweets, columns=['id', 'handle', 'created_at', 'text'])
    return tweets_df

# Extract tweets
api = create_api(consumer_key, consumer_key_secret)
search_results = search_tweets(api, 'apple', ignore_rt=True, max_tweets=1000)
tweets = search_to_df(search_results)
tweets.head()

# Export the tweets into CSV
tweets.to_csv('out/tweets.csv', index=False)
